refactor main.rb and app.rb to adhere to SOLID principles
